add stack management to listing

+++ DTOs/ShowStackableDTO.cs
@@ -4,7 +4,7 @@ namespace InfiniteCreativity.DTO
 {
     public class ShowStackableDTO : ShowItemDTO
     {
-        public int Amount { get; set; }
+        public ulong Amount { get; set; }
         public int StackSize { get; set; }
         public StackableType StackableType { get; set; }
     }
